switch parallesization to misc.parallelize and run_cmd to shell

+++ src/aztools/data.py
@@ -1,14 +1,12 @@
 """Data Tools"""
 
-import functools
 import glob
 import os
-from multiprocessing import Pool
 
 import numpy as np
 from astropy.io import fits
 
-from .misc import hsp, run_cmd_line_tool
+from .misc import hsp, parallelize, run_cmd_line_tool
 
 __all__ = [
     'process_nicer_obsid', 'process_nicer_obsids',
@@ -22,48 +20,6 @@ __all__ = [
     'extract_nustar_lc', 'extract_nustar_lcs',
     'extract_nicer_spec', 'extract_nicer_specs'
 ]
-
-
-def _make_parallel(func, nproc=4):
-    """A wrapper to make a function run in parallel
-    
-    Parameters
-    ----------
-    func: method
-        The method to parallelize. It should expects on obsid
-    nproc: int
-        Number of processes to run
-
-    Return
-    ------
-    return a method that takes a list of obsids and calls func
-    on each of them in parallel
-    
-    """
-
-    @functools.wraps(func)
-    def parallelize(obsids, **kwargs):
-
-        if isinstance(obsids, str):
-            obsids = [obsids]
-
-        irun = kwargs.pop('irun', None)
-        if irun is not None:
-            if isinstance(irun, int):
-                iruns = list(range(irun, irun+len(obsids)))
-            elif isinstance(irun, (list, np.ndarray)) and len(irun) == len(obsids):
-                iruns = irun
-            else:
-                raise ValueError(f'{irun} should be either an int or an array '
-                                 'of length similar to obsids')
-        obsids = [f'{o}:{ir}' for o,ir in zip(obsids, iruns)]
-
-        with Pool(min(nproc, len(obsids))) as pool:
-            results = pool.map(functools.partial(func, **kwargs), obsids)
-
-        return results
-
-    return parallelize
 
 
 def process_nicer_obsid(obsid: str, **kwargs):
@@ -117,7 +73,7 @@ def process_nicer_obsid(obsid: str, **kwargs):
 
 
 # parallel version of process_nicer_obsid
-process_nicer_obsids = _make_parallel(process_nicer_obsid)
+process_nicer_obsids = parallelize(process_nicer_obsid, use_irun=False)
 
 
 def process_nustar_obsid(obsid: str, **kwargs):
@@ -174,7 +130,7 @@ def process_nustar_obsid(obsid: str, **kwargs):
     return out.returncode
 
 # parallel version of process_nicer_obsid
-process_nustar_obsids = _make_parallel(process_nustar_obsid)
+process_nustar_obsids = parallelize(process_nustar_obsid, use_irun=False)
 
 
 def process_suzaku_obsid(obsid: str, **kwargs):
@@ -232,7 +188,17 @@ def process_suzaku_obsid(obsid: str, **kwargs):
     return out.returncode
 
 # parallel version of process_suzaku_obsid
-process_suzaku_obsids = _make_parallel(process_suzaku_obsid)
+process_suzaku_obsids = parallelize(process_suzaku_obsid, use_irun=False)
+
+
+def _run_sas_cmd(cmd, *args, **kwargs):
+    """prefix run_cmd_line_tool with sas initialization"""
+
+    if 'SAS_DIR' not in os.environ:
+        raise KeyError('SAS_DIR is not defined')
+
+    pre_cmd = 'source $SAS_DIR/sas-setup.sh; '
+    return run_cmd_line_tool(pre_cmd + cmd, *args, **kwargs)
 
 
 def process_xmm_obsid(obsid: str, **kwargs):
@@ -249,10 +215,14 @@ def process_xmm_obsid(obsid: str, **kwargs):
     --------
     instr: str
         Instrument or instrument mode: pn|mos|rgs|om
+    odfingest: bool
+        run odfingest? This should be True for the first run
+        and False subsequently
     Any parameters to be passed to the reduction pipeline
     
     
     """
+    odfingest = kwargs.pop('odfingest', True)
 
     # defaults
     instr = kwargs.pop('instr', 'pn')
@@ -272,14 +242,15 @@ def process_xmm_obsid(obsid: str, **kwargs):
             if len(glob.glob('*gz')) > 0:
                 os.system('gzip -d *gz')
             cmd = 'cifbuild withccfpath=no analysisdate=now category=XMMCCF fullpath=yes'
-            run_cmd_line_tool(cmd, env, logfile='processing_xmm_cifbuild.log')
+            _run_sas_cmd(cmd, env, logfile='processing_xmm_cifbuild.log')
 
         env['SAS_CCF'] = f'{os.getcwd()}/ccf.cif'
 
-        if len(glob.glob('*.SAS')) > 0:
+        if len(glob.glob('*.SAS')) > 0 and odfingest:
             os.system('rm *.SAS')
-        cmd = f'odfingest odfdir={os.getcwd()} outdir={os.getcwd()}'
-        run_cmd_line_tool(cmd, env, logfile='processing_xmm_odfingest.log')
+        if odfingest:
+            cmd = f'odfingest odfdir={os.getcwd()} outdir={os.getcwd()}'
+            _run_sas_cmd(cmd, env, logfile='processing_xmm_odfingest.log')
 
 
         # prepare the command
@@ -297,9 +268,9 @@ def process_xmm_obsid(obsid: str, **kwargs):
         else:
             raise ValueError('instr needs to be pn|om|rgs|om')
 
-        # the following with raise RuntimeError if the task fails
+        # the following will raise RuntimeError if the task fails
         cmd = f'{cmd} {" ".join([f"{par}={val}" for par,val in kwargs.items()])}'
-        run_cmd_line_tool(cmd, env, logfile=f'processing_xmm_{instr}.log')
+        _run_sas_cmd(cmd, env, logfile=f'processing_xmm_{instr}.log')
 
         # post run extra tasks
         if instr == 'pn':
@@ -328,7 +299,7 @@ def process_xmm_obsid(obsid: str, **kwargs):
             cmd = ('rgscombine pha="spec_r1.src spec_r2.src" bkg="spec_r1.bgd spec_r2.bgd" '
                    'rmf="spec_r1.rsp spec_r2.rsp" filepha="spec_rgs.src" filebkg="spec_rgs.bgd" '
                    'filermf="spec_rgs.rsp"')
-            run_cmd_line_tool(cmd, env, logfile='processing_xmm_rgscombine.log')
+            _run_sas_cmd(cmd, env, logfile='processing_xmm_rgscombine.log')
             print('rgs spectra created successfully')
 
         os.chdir(cwd)
@@ -338,7 +309,7 @@ def process_xmm_obsid(obsid: str, **kwargs):
         raise exception
 
 # parallel version of process_xmm_obsid
-process_xmm_obsids = _make_parallel(process_xmm_obsid)
+process_xmm_obsids = parallelize(process_xmm_obsid, use_irun=False)
 
 
 def filter_xmm_obsid(obsid: str, **kwargs):
@@ -404,33 +375,33 @@ def filter_xmm_obsid(obsid: str, **kwargs):
             f"#XMMEA_{foptions[instr[:3]][1]}' "
             "makeratecolumn=yes maketimecolumn=yes timebinsize=100"
         )
-        run_cmd_line_tool(cmd, logfile='filter_xmm_evselect.log')
+        _run_sas_cmd(cmd, logfile='filter_xmm_evselect.log')
 
         # generate GTI
         expr = gtiexpr or f'RATE < {foptions[instr[:3]][3]}'
         cmd = ('tabgtigen table=tmp.rate gtiset=tmp.gti '
                f'expression="RATE<{foptions[instr[:3]][3]}"')
-        run_cmd_line_tool(cmd, logfile='filter_xmm_tabgtigen.log')
+        _run_sas_cmd(cmd, logfile='filter_xmm_tabgtigen.log')
 
         # apply GTI
         expr = ('gti(tmp.gti,TIME)&&(PI IN [200:12000])&&(FLAG==0)&&'
                 f'(PATTERN<={foptions[instr[:3]][2]}){extra_expr}')
         cmd = (f"evselect table={evt}:EVENTS withfilteredset=yes "
                f"filteredset={filtered_evt} expression='{expr}'")
-        run_cmd_line_tool(cmd, logfile='filter_xmm_evselect2.log')
+        _run_sas_cmd(cmd, logfile='filter_xmm_evselect2.log')
 
         # barycenter corrections? #
         if barycorr:
             os.system(f'cp {filtered_evt} {instr}_filtered_nobary.fits'.format(instr))
             cmd = f'barycen table={filtered_evt}:EVENTS'
-            run_cmd_line_tool(cmd, logfile='filter_xmm_barycen.log')
+            _run_sas_cmd(cmd, logfile='filter_xmm_barycen.log')
 
         # region?
         if region:
             # create an image
             cmd = (f"evselect table={filtered_evt}:EVENTS withimageset=yes "
                    "imageset=tmp.img xcolumn=X ycolumn=Y")
-            run_cmd_line_tool(cmd, logfile='filter_xmm_image.log')
+            _run_sas_cmd(cmd, logfile='filter_xmm_image.log')
 
             # call ds9
             print('launching ds9')
@@ -446,7 +417,7 @@ def filter_xmm_obsid(obsid: str, **kwargs):
         raise exception
 
 # parallel version of filter_xmm_obsid
-filter_xmm_obsids = _make_parallel(filter_xmm_obsid)
+filter_xmm_obsids = parallelize(filter_xmm_obsid, use_irun=False)
 
 
 def extract_xmm_spec(obsid: str, **kwargs):
@@ -473,6 +444,8 @@ def extract_xmm_spec(obsid: str, **kwargs):
         Note the &&.
     genrsp: bool
         Generate rmf and arf files.
+    irun: int
+        name suffix, so the output is spec_{irun}*
     
     """
     # get keywords
@@ -490,8 +463,8 @@ def extract_xmm_spec(obsid: str, **kwargs):
                           'to connect to other expression'))
 
     prefix = 'spec'
-    if ':' in obsid:
-        obsid, irun = obsid.split(':')
+    irun = kwargs.get('irun', None)
+    if irun is not None:
         prefix += f'_{irun}'
     cwd = os.getcwd()
 
@@ -531,26 +504,26 @@ def extract_xmm_spec(obsid: str, **kwargs):
                    'energycolumn=PI spectralbinsize=5 withspecranges=yes '
                    f'specchannelmin=0 specchannelmax={maxchan}'
             )
-            run_cmd_line_tool(cmd, env, logfile=f'extract_xmm_spec_{lab}.log')
+            _run_sas_cmd(cmd, env, logfile=f'extract_xmm_spec_{lab}.log')
             # backscale
             cmd = f'backscale spectrumset={prefix}.{lab} badpixlocation={filtered_evt}'
-            run_cmd_line_tool(cmd, env, logfile=f'extract_xmm_spec_{lab}_backscale.log')
+            _run_sas_cmd(cmd, env, logfile=f'extract_xmm_spec_{lab}_backscale.log')
 
         if genrsp:
             # response
             cmd = f'rmfgen spectrumset={prefix}.pha rmfset={prefix}.rmf'
-            run_cmd_line_tool(cmd, env, logfile='extract_xmm_spec_rmf.log')
+            _run_sas_cmd(cmd, env, logfile='extract_xmm_spec_rmf.log')
 
             # arf
             cmd = (f'arfgen spectrumset={prefix}.pha arfset={prefix}.arf withrmfset=yes '
                    f'rmfset={prefix}.rmf badpixlocation={filtered_evt} detmaptype=psf')
-            run_cmd_line_tool(cmd, env, logfile='extract_xmm_spec_arf.log')
+            _run_sas_cmd(cmd, env, logfile='extract_xmm_spec_arf.log')
 
             # group spectra
             cmd = (f'specgroup spectrumset={prefix}.pha rmfset={prefix}.rmf '
                    f'backgndset={prefix}.bgd arfset={prefix}.arf groupedset={prefix}.grp '
                    'minSN=6 oversample=3')
-            run_cmd_line_tool(cmd, env, logfile='extract_xmm_spec_grp.log')
+            _run_sas_cmd(cmd, env, logfile='extract_xmm_spec_grp.log')
 
         os.chdir(cwd)
         print(f'spectra {prefix}* created successfully')
@@ -560,7 +533,7 @@ def extract_xmm_spec(obsid: str, **kwargs):
         raise exception
 
 # parallel version of extract_xmm_spec
-extract_xmm_specs = _make_parallel(extract_xmm_spec)
+extract_xmm_specs = parallelize(extract_xmm_spec, use_irun=True)
 
 
 def extract_xmm_lc(obsid: str, **kwargs):
@@ -593,6 +566,8 @@ def extract_xmm_lc(obsid: str, **kwargs):
         Run epiclccorr. Default is True.
     outdir: str
         output folder name under {obsid}/{instr}/. Default is lc.
+    irun: int
+        name suffix so the output file is lc_{irun}*
     
     """
     # get keywords
@@ -603,6 +578,7 @@ def extract_xmm_lc(obsid: str, **kwargs):
     extra_expr = kwargs.pop('extra_expr', '')
     lccorr = kwargs.pop('lccorr', True)
     outdir = kwargs.pop('outdir', 'lc')
+    irun = kwargs.get('irun', None)
 
     # check keywords
     if instr not in ['pn', 'mos1', 'mos2']:
@@ -619,8 +595,7 @@ def extract_xmm_lc(obsid: str, **kwargs):
                           'to connect to other expression'))
 
     prefix = f'lc_{tbin:03g}'
-    if ':' in obsid:
-        obsid, irun = obsid.split(':')
+    if irun is not None:
         prefix += f'_{irun}'
     cwd = os.getcwd()
 
@@ -663,7 +638,7 @@ def extract_xmm_lc(obsid: str, **kwargs):
             cmd = (f'evselect table={filtered_evt}:EVENTS withrateset=yes '
                    f'rateset={src} expression="{expr}" makeratecolumn=yes '
                    f'maketimecolumn=yes timebinsize={tbin}')
-            run_cmd_line_tool(cmd, env, logfile='extract_xmm_lc_src.log')
+            _run_sas_cmd(cmd, env, logfile='extract_xmm_lc_src.log')
 
 
             # extract bgd lc
@@ -671,14 +646,14 @@ def extract_xmm_lc(obsid: str, **kwargs):
             cmd = (f'evselect table={filtered_evt}:EVENTS withrateset=yes '
                    f'rateset={bgd} expression="{expr}" makeratecolumn=yes '
                    f'maketimecolumn=yes timebinsize={tbin}')
-            run_cmd_line_tool(cmd, env, logfile='extract_xmm_lc_bgd.log')
+            _run_sas_cmd(cmd, env, logfile='extract_xmm_lc_bgd.log')
 
 
             # correct lc
             if lccorr:
                 cmd = (f'epiclccorr srctslist={src} eventlist={filtered_evt} outset={smb} '
                        f'withbkgset=yes bkgtslist={bgd} applyabsolutecorrections=no')
-                run_cmd_line_tool(cmd, env, logfile='extract_xmm_lc_corr.log')
+                _run_sas_cmd(cmd, env, logfile='extract_xmm_lc_corr.log')
 
         os.chdir(cwd)
         print(f'Light cruves {prefix}* created successfully')
@@ -688,7 +663,7 @@ def extract_xmm_lc(obsid: str, **kwargs):
         raise exception
 
 # parallel version of extract_xmm_lc
-extract_xmm_lcs = _make_parallel(extract_xmm_lc)
+extract_xmm_lcs = parallelize(extract_xmm_lc, use_irun=True)
 
 
 def extract_nustar_spec(obsid: str, **kwargs):
@@ -706,6 +681,8 @@ def extract_nustar_spec(obsid: str, **kwargs):
     --------
     processed_obsid: str
         The name of the processed obsid folder. Default: {obsid}_p
+    irun: int
+        name suffix so the output is spec_{irun}*
     
     Any parameters to be passed to the reduction pipeline
     
@@ -718,10 +695,10 @@ def extract_nustar_spec(obsid: str, **kwargs):
         raise ImportError('extract_nustar_spec depends on heasoftpy. Install it first')
 
     processed_obsid = kwargs.pop('processed_obsid', None)
+    irun = kwargs.pop('irun', None)
 
     prefix = 'spec'
-    if ':' in obsid:
-        obsid, irun = obsid.split(':')
+    if irun is not None:
         prefix += f'_{irun}'
     if processed_obsid is None:
         processed_obsid = f'{obsid}_p'
@@ -798,7 +775,7 @@ def extract_nustar_spec(obsid: str, **kwargs):
     return 0
 
 # parallel version of extract_nustar_spec
-extract_nustar_specs = _make_parallel(extract_nustar_spec)
+extract_nustar_specs = parallelize(extract_nustar_spec, use_irun=True)
 
 
 def extract_nustar_lc(obsid: str, **kwargs):
@@ -826,6 +803,8 @@ def extract_nustar_lc(obsid: str, **kwargs):
         Barycenter the light curve (assume gti if used is bary-centered). Default is False.
     outdir: str
         output folder name under {processed_obsid}/. Default is lc.
+    irun: int or str
+        name suffix so that the output is lc_{irun}*
     
     
     Any parameters to be passed to the reduction nuproduct. e.g usrgtifile="somefile.gti"
@@ -844,10 +823,11 @@ def extract_nustar_lc(obsid: str, **kwargs):
     lccorr = kwargs.pop('lccorr', True)
     barycorr = kwargs.pop('barycorr', False)
     outdir = kwargs.pop('outdir', 'lc')
+    irun = kwargs.get('irun', None)
+
 
     prefix = 'lc'
-    if ':' in obsid:
-        obsid, irun = obsid.split(':')
+    if irun is not None:
         prefix += f'_{irun}'
     if processed_obsid is None:
         processed_obsid = f'{obsid}_p'
@@ -966,7 +946,7 @@ def extract_nustar_lc(obsid: str, **kwargs):
     return 0
 
 # parallel version of extract_nustar_lc
-extract_nustar_lcs = _make_parallel(extract_nustar_lc)
+extract_nustar_lcs = parallelize(extract_nustar_lc, use_irun=True)
 
 
 def extract_nicer_spec(obsid: str, **kwargs):
@@ -982,6 +962,8 @@ def extract_nicer_spec(obsid: str, **kwargs):
     
     Keywords
     --------
+    irun: int or str
+        name suffix so the output is spec_{irun}*
     Any parameters to be passed to the reduction pipeline
     
     Return
@@ -992,10 +974,9 @@ def extract_nicer_spec(obsid: str, **kwargs):
     if hsp is None:
         raise ImportError('extract_nicer_spec depends on heasoftpy. Install it first')
 
-
+    irun = kwargs.get('irun', None)
     prefix = 'spec'
-    if ':' in obsid:
-        obsid, irun = obsid.split(':')
+    if irun is not None:
         prefix += f'_{irun}'
     processed_obsid = obsid
 
@@ -1060,4 +1041,4 @@ def extract_nicer_spec(obsid: str, **kwargs):
     return 0
 
 # parallel version of extract_nustar_lc
-extract_nicer_specs = _make_parallel(extract_nicer_spec)
+extract_nicer_specs = parallelize(extract_nicer_spec, use_irun=True)
